Reorganize aries_vcx_core files and code
