+++ aries_vcx_core/src/anoncreds/indy/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod proofs;
+pub(crate) mod credentials;
+pub(crate) mod general;
+pub mod primitives;
