+++ aries_vcx_core/src/anoncreds/indy/primitives/mod.rs
@@ -1,0 +1,1 @@
+pub mod credential_schema;
