+++ aries_vcx_core/src/anoncreds/indy/proofs/mod.rs
@@ -1,0 +1,2 @@
+pub mod prover;
+pub mod verifier;
