+++ aries_vcx_core/src/anoncreds/mod.rs
@@ -3,3 +3,4 @@ pub mod base_anoncreds;
 pub mod credx_anoncreds;
 #[cfg(feature = "vdrtools")]
 pub mod indy_anoncreds;
+pub mod indy;
