--- aries_vcx_core/src/common/ledger/mod.rs
@@ -1,1 +1,0 @@
-pub mod transactions;
