--- aries_vcx_core/src/common/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod ledger;
