+++ aries_vcx_core/src/indy/mod.rs
@@ -1,7 +1,2 @@
-// TODO - FUTURE - visibility of all indy should be 'crate' to confirm indy dependency is stripped
-pub(crate) mod anoncreds;
-pub(crate) mod credentials;
-pub mod ledger; // temporarily left public due to pool set up utils
-pub(crate) mod primitives;
-pub(crate) mod proofs;
+pub mod ledger;
 pub mod utils;
