--- aries_vcx_core/src/indy/primitives/mod.rs
@@ -1,3 +1,0 @@
-pub mod credential_definition;
-pub mod credential_schema;
-pub mod revocation_registry;
