--- aries_vcx_core/src/indy/proofs/mod.rs
@@ -1,2 +1,0 @@
-pub mod prover;
-pub mod verifier; // TODO: Should stay private
