+++ aries_vcx_core/src/ledger/common.rs
@@ -6,51 +6,12 @@ use crate::errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind, VcxCoreResu
 
 #[derive(Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
-pub struct Request {
+struct Request {
     pub req_id: u64,
     pub identifier: String,
     pub signature: Option<String>,
     pub signatures: Option<HashMap<String, String>>,
     pub endorser: Option<String>,
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(tag = "op")]
-pub enum Response {
-    #[serde(rename = "REQNACK")]
-    ReqNACK(Reject),
-    #[serde(rename = "REJECT")]
-    Reject(Reject),
-    #[serde(rename = "REPLY")]
-    Reply(Reply),
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct Reject {
-    pub reason: String,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-pub enum Reply {
-    ReplyV0(ReplyV0),
-    ReplyV1(ReplyV1),
-}
-
-#[derive(Debug, Deserialize)]
-pub struct ReplyV0 {
-    pub result: serde_json::Value,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct ReplyV1 {
-    pub data: ReplyDataV1,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct ReplyDataV1 {
-    pub result: serde_json::Value,
 }
 
 pub fn verify_transaction_can_be_endorsed(transaction_json: &str, submitter_did: &str) -> VcxCoreResult<()> {
@@ -74,10 +35,10 @@ pub fn verify_transaction_can_be_endorsed(transaction_json: &str, submitter_did:
     let identifier = transaction.identifier.as_str();
     if transaction.signature.is_none()
         && !transaction
-            .signatures
-            .as_ref()
-            .map(|signatures| signatures.contains_key(identifier))
-            .unwrap_or(false)
+        .signatures
+        .as_ref()
+        .map(|signatures| signatures.contains_key(identifier))
+        .unwrap_or(false)
     {
         return Err(AriesVcxCoreError::from_msg(
             AriesVcxCoreErrorKind::InvalidJson,
